Fixed Error Handling + Added getpass
Any invalid login information will now output "Error with login, please try again." It does this by seeing whether there is null, and if it does, then it will know that something went wrong. Also added getpass, which hides the user input; in this case, hiding the password typed in.