Added Rudimentary GUI
Using tkinter (a built-in library dealing with basic GUIs), I made the function more "interface-y". Now there's entry boxes for username and password, a button to get the balance, and a message box that outputs the balance (plus some errors the user may have made). It looks basic, but it works; I'll create a better design after the core functionality of scanning and getting are functional.

Message me on Slack once you've tried it yourself.